refactor(seed): separate quiz tags data from logic
